+++ src/components/TodoModal/index.ts
@@ -1,0 +1,5 @@
+export { TodoModalDocs } from './TodoModalDocs';
+export { TodoModalHeader } from './TodoModalHeader';
+export { TodoModalTarget } from './TodoModalTarget';
+export { TodoModalTitle } from './TodoModalTitle';
+export { TodoModalRepeat } from './TodoModalRepeat';
